added Generics to Gamer and Jogos DAO

+++ src/main/kotlin/br/com/alura/alugames/modelo/Plano.kt
@@ -1,9 +1,7 @@
 package br.com.alura.alugames.modelo
 
-import java.math.BigDecimal
-
 sealed class Plano(val tipo: String, var id: Int = 0) {
-    open fun obterValor(aluguel: Aluguel): BigDecimal {
-        return aluguel.jogo.preco * aluguel.periodo.emDias.toBigDecimal()
+    open fun obterValor(aluguel: Aluguel): Double {
+        return aluguel.jogo.preco * aluguel.periodo.emDias
     }
 }
